+++ Client/src/pages/installed-video-editor.aspx.js
@@ -45,22 +45,6 @@ class installedVideoEditor extends React.PureComponent {
   }
 
   componentDidMount(){
-    const queryString = require('query-string');
-    const parsed = queryString.parse(document.location.search);
-    const cookies = new Cookies();
-    if (parsed.SRC) {
-      cookies.set('SRC', parsed.SRC, { path: '/' });
-    }
-  
-    const SRCParam = cookies.get('SRC')
-  
-    if(SRCParam){
-  
-      this.setState({
-        hrefUnlim: this.state.hrefUnlim+"&SRC="+SRCParam,
-      })
-    }
-  
     this.setState({
      documentLoaded: true
    })
@@ -87,9 +71,9 @@ render(){
           </div>
           <ImageGQL className="headerBackgroundImage" imageName="installed-video-editor-header.jpg" style={{position: "absolute"}}/>
             <div className="header__body">  <ImageGQL className="installedSuper" imageName="installed-video-editor-super.png" style={{margin: "auto"}}/>
-              <Text as="h1" className="header__heading installed">{this.props.t("Thank you for installing AVS Video Editor")}</Text>
+              <Text as="h1" className="header__heading installed">{this.props.t("Thank you for installing AVS Video Editor")} <br /> {this.props.t("AVS video editor")}</Text>
              <table className="header__heading exclusive"><Text as="h2" className="header__heading exclusive"><a href={this.props.t(`${this.state.hrefUnlim}`)} style={{color: "#fff"}}>{this.props.t("Exclusive Offer Only Today")}</a></Text></table>
-              <Text className="header__subtitle installed">{this.props.t("Save 70 on the full version and edit your audio without limitations")}</Text>
+              <Text className="header__subtitle installed">{this.props.t("Save 70 on the full version and edit your video without limitations")}</Text>
               <table className="header__price__block"><Text as="h4" className="header__price">{this.props.t("199")}</Text>
               <Text as="h4" className="header__new__price">{this.props.t("590")}</Text></table>
               <table className="header__buy"><Text as="h2" className="header__buy__now"><a href={this.props.t(`${this.state.hrefUnlim}`)} style={{color: "#fff"}}>{this.props.t("Buy now")}</a></Text></table>
@@ -122,7 +106,7 @@ render(){
         <div className="contact-body">
           <div className="contact-body__max-width">
             <Text as="h2" className="common__heading">
-            + {this.props.t("Get 4 professional multimedia")}
+            + {this.props.t("Get 4 professional multimedia programs as a gift")}
             </Text>
             <div className="contact-body__wrapper">
 
@@ -130,7 +114,7 @@ render(){
                 <div className="gift">{this.props.t("Gift")}</div>
                 <div className="download-product"><a href="https://downloads.avs4you.com/distributives/AVSAudioEditor.exe" style={{color: "#fff"}}>{this.props.t("Download now")}</a></div>
                 <Text as="h4" className="common-info-block__headline">
-                AVS Audio Editor
+                {this.props.t("AVS Audio Editor name")}
                 </Text>
                 <Text className="common-info-block__text">
                 {this.props.t("Professional level audio editing and recording software for Windows")}
@@ -141,7 +125,7 @@ render(){
               <div className="gift">{this.props.t("Gift")}</div>
               <div className="download-product"><a href="https://downloads.avs4you.com/distributives/AVSVideoConverter.exe" style={{color: "#fff"}}>{this.props.t("Download now")}</a></div>
                 <Text as="h4" className="common-info-block__headline">
-                AVS Video Converter
+                {this.props.t("AVS Video Converter name")}
                 </Text>
                 <Text className="common-info-block__text">
                 {this.props.t("Allinone video converter with basic video editing features")}
@@ -152,7 +136,7 @@ render(){
               <div className="gift">{this.props.t("Gift")}</div>
               <div className="download-product"><a href="https://downloads.avs4you.com/distributives/AVSAudioConverter.exe" style={{color: "#fff"}}>{this.props.t("Download now")}</a></div>
                 <Text as="h4" className="common-info-block__headline">
-                AVS Audio Converter
+                {this.props.t("AVS Audio Converter name")}
                 </Text>
                 <Text className="common-info-block__text">
                 {this.props.t("A multiformat audio conversion application with basic audio editing tools")}
@@ -163,7 +147,7 @@ render(){
               <div className="gift">{this.props.t("Gift")}</div>
               <div className="download-product"><a href="https://downloads.avs4you.com/distributives/AVSVideoReMaker.exe" style={{color: "#fff"}}>{this.props.t("Download now")}</a></div>
                 <Text as="h4" className="common-info-block__headline">
-                AVS Video ReMaker
+                {this.props.t("AVS Video ReMaker name")}
                 </Text>
                 <Text className="common-info-block__text">
                 {this.props.t("A fast video editing tool to make edits without changing the original video file format")}